+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Notable changes to the `alacritty_terminal` crate are documented in its
 
 ## 0.14.0-dev
 
+### Added
+
+- New `strength` factor and setting for the visual bell effect
+
 ### Changed
 
 - Pressing `Alt` with unicode input will now add `ESC` like for ASCII input